+++ internal/provider/resource_environment.go
@@ -191,34 +191,6 @@ func environmentImport(ctx context.Context, d *schema.ResourceData, meta interfa
 	return []*schema.ResourceData{d}, nil
 }
 
-func environmentImporter() *Importer {
-	return &Importer{
-		LoadInstanceIds: loadAllEnvironments,
-	}
-}
-
-func loadAllEnvironments(ctx context.Context, client *Client) (InstanceIdsToNameMap, diag.Diagnostics) {
-	instances := make(InstanceIdsToNameMap)
-
-	environments, err := loadEnvironments(ctx, client)
-	if err != nil {
-		tflog.Warn(ctx, fmt.Sprintf("Error reading Environments: %s", createDescriptiveError(err)))
-		return instances, diag.FromErr(createDescriptiveError(err))
-	}
-	environmentsJson, err := json.Marshal(environments)
-	if err != nil {
-		return instances, diag.Errorf("error reading Environments: error marshaling %#v to json: %s", environments, createDescriptiveError(err))
-	}
-	tflog.Debug(ctx, fmt.Sprintf("Fetched Environments: %s", environmentsJson))
-
-	for _, environment := range environments {
-		instanceId := environment.GetId()
-		instances[instanceId] = toValidTerraformResourceName(environment.GetDisplayName())
-	}
-
-	return instances, nil
-}
-
 func getNestedStreamGovernancePackageKey() string {
 	return fmt.Sprintf("%s.0.%s", paramStreamGovernance, paramPackage)
 }
